cargo miri test
Singling out a single test to run under miri actually increases the load on CI as it has to construct two separate environments and build the crate twice.

Now that this is a standalone crate, it can simply run all of its tests under miri without affecting the testing of other crates.